Updte both y axis to PL values
